+++ backend/main_test.go
@@ -1,3 +1,13 @@
+// Copyright 2021 The Elabox Authors
+// This file is part of elabox-logs library.
+
+// elabox-logs library is under open source LGPL license.
+// If you simply compile or link an LGPL-licensed library with your own code,
+// you can release your application under any license you want, even a proprietary license.
+// But if you modify the library or copy parts of it into your code,
+// you’ll have to release your application under similar terms as the LGPL.
+// Please check license description @ https://www.gnu.org/licenses/lgpl-3.0.txt
+
 package main
 
 import (
@@ -5,6 +15,7 @@ import (
 	"testing"
 )
 
+// test log retrieval
 func TestRetrieveLogs(t *testing.T) {
 	res := RetrieveLatestOffset()
 	output, err := json.Marshal(res)
@@ -16,6 +27,7 @@ func TestRetrieveLogs(t *testing.T) {
 	t.Log(output)
 }
 
+// test filter retrieval
 func TestRetrieveFilter(t *testing.T) {
 	summary := LoadLogSummary()
 	json, err := json.Marshal(summary)
