+++ react_client/src/constant.js
@@ -1,3 +1,13 @@
+// Copyright 2021 The Elabox Authors
+// This file is part of elabox-logs library.
+
+// elabox-logs library is under open source LGPL license.
+// If you simply compile or link an LGPL-licensed library with your own code,
+// you can release your application under any license you want, even a proprietary license.
+// But if you modify the library or copy parts of it into your code,
+// you’ll have to release your application under similar terms as the LGPL.
+// Please check license description @ https://www.gnu.org/licenses/lgpl-3.0.txt
+
 export const PKID = "ela.logs"
 export const LOAD_FILTERS_AC = PKID + ".LOAD_FILTERS_ACTION"  // use to populate data for filters
 export const LOAD_LATEST_AC = PKID + ".LOAD_LATEST_ACTION"    // use to load latest log action
